pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "de-dv-airflow-local-pro"
version = "1.0.0"
description = "Comprehensive Data Engineering Platform with Streaming, Batch, Pipelines and Airflow 2.11"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "DE/DV Team" }]
keywords = ["airflow", "kafka", "spark", "streaming", "batch", "etl", "data-engineering"]

dependencies = [
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.1",
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "requests>=2.32.3",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
streaming = [
    "kafka-python>=2.0.2",
    "confluent-kafka>=2.4.0",
    "fastavro>=1.9.4",
    "pyspark>=3.5.1",
]
batch = [
    "pyspark>=3.5.1",
    "pandas>=2.2.2",
    "SQLAlchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "pyarrow>=16.0.0",
]
airflow = [
    "apache-airflow>=2.11.0",
    "apache-airflow-providers-apache-spark>=4.9.0",
    "apache-airflow-providers-apache-kafka>=1.4.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "moto[s3,sqs]>=5.0.7",
]
all = [
    "de-dv-airflow-local-pro[streaming,batch,airflow,dev]",
]

[tool.setuptools.packages.find]
include = ["streaming_apps*", "batch_apps*", "pipelines*", "airflow*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing -v"

[tool.ruff]
line-length = 100
target-version = "py311"
